Updated delete booking method.

+++ sas/booking/views.py
@@ -163,10 +163,16 @@ def delete_booking(request, id):
 		booking = Booking.objects.get(pk=id)
 		booktimes = booking.time.all()
 		for booktime in booktimes:
-			booktime.delete()
+			if not booktime.booking_time.all().exclude(pk = id).exists():
+				booktime.delete()
 		booking.delete()
 		messages.success(request, _('Booking deleted!'), {})
 		return render(request, 'booking/searchBooking.html', {})
 	else:
 		messages.error(request, _('You cannot delete this booking.'))
 		return render(request, 'booking/searchBooking.html', {})
+
+def delete_specific_booktime(request, id):
+	if has_permission_to_delete(request):
+		id = int(id)
+		BookTime.objects.get(pk = id).delete()
